requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
readability-lxml>=0.8.0
html2text>=2020.0
beautifulsoup4>=4.12.0
requests>=2.28.0
